refactor: better imports and folder structure 📁
