+++ src/components/Navbar/Navbar.tsx
@@ -1,6 +1,6 @@
 import { Box, chakra, Flex, HStack, Spacer } from '@chakra-ui/react'
 import { type ReactElement } from 'react'
-import { useCustomTheme } from '~/hooks/useCustomTheme/useCustomTheme'
+import { useCustomTheme } from '~/hooks/useCustomTheme'
 import ThemeChanger from '~/components/Navbar/ThemeChanger'
 import MainMenu from '~/components/MainMenu/MainMenu'
